fix(swissprot data): ensure that mapping from annot to index includes all annots present in samples
When running the training procedure, I ran into a problem where the annotation of the sample from
the loader was not present in the mapper (trg2idx). This was because the way we setup the mapper
assumed we use get_samples method. However, we use get_samples_ic, therefore I adjusted the
construction of encodings by removing the filtering of annotations based on their level which
ensures that now all the annotations we have in the data can be found also in the mapper.